+++ CHANGELOG.md
@@ -1066,3 +1066,4 @@
 - [2026-05-12 12:14] docs(theme): update readme and API contracts for theme
 - [2026-05-12 12:59] fix(network): prevent double tap duplicate action in network
 - [2026-05-12 14:32] fix(test): fix memory leak by cancelling test subscription
+- [2026-05-12 14:49] refactor(security): decouple repository interface from concrete security service